% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HyperGen.R
\name{Add}
\alias{Add}
\title{Custom Addition Function}
\usage{
Add(x, y)
}
\arguments{
\item{x}{An object of class `GFElem` or `PGInf`.}

\item{y}{An object of class `GFElem` or `PGInf`.}
}
\value{
An object of class `GFElem` if both arguments are `GFElem`, or an object of class `PGInf` if either argument is `PGInf`.
}
\description{
Adds two elements from the Galois Field or one element from the Galois Field and one element representing infinity.
}
\details{
This function performs addition for elements of the Galois Field (GFElem) or between a Galois Field element and a point at infinity (PGInf).
If both elements are of type `GFElem`, their values are added modulo the field's characteristic. If either element is of type `PGInf`, the result is a point at infinity.
}
\examples{
# Create two Galois Field elements
elem1 <- GFElem$new(3, 7)
elem2 <- GFElem$new(5, 7)

# Add the two Galois Field elements
result <- Add(elem1, elem2)
result$show() # Should display the result in the format GFElem{7}(1)

# Create a Galois Field element and a point at infinity
elem3 <- GFElem$new(2, 7)
inf <- PGInf$new(7)

# Add the Galois Field element and the point at infinity
result2 <- Add(elem3, inf)
result2$show() # Should display PGInf{7}()

}
