% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GCDEEA}
\alias{GCDEEA}
\title{Greatest Common Divisor and Extended Euclidean Algorithm}
\usage{
GCDEEA(a, b)
}
\arguments{
\item{a}{An integer. The first number for which the GCD is to be computed.}

\item{b}{An integer. The second number for which the GCD is to be computed.}
}
\value{
A list with three elements:
\describe{
  \item{\eqn{g}}{The greatest common divisor of \eqn{a} and \eqn{b}.}
  \item{\eqn{s}}{The coefficient of \eqn{a} in the linear combination that equals \eqn{g}.}
  \item{\eqn{t}}{The coefficient of \eqn{b} in the linear combination that equals \eqn{g}.}
}
}
\description{
Computes the greatest common divisor (GCD) of two integers using the Extended Euclidean Algorithm.
This algorithm not only computes the GCD but also finds coefficients \eqn{s} and \eqn{t} such that:
\deqn{g = a \cdot s + b \cdot t}
where \eqn{g} is the GCD of \eqn{a}  and \eqn{b}.
}
\examples{
# Compute the greatest common divisor and the coefficients for 30 and 12
result <- GCDEEA(30, 12)
print(result)

# Expected output: A list with the GCD, and coefficients s and t
# Example output: $g
# [1] 6
# $s
# [1] 1
# $t
# [1] -2
}
