% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Gen2Factors}
\alias{Gen2Factors}
\title{Generate Factors for \eqn{q = 2}}
\usage{
Gen2Factors(m)
}
\arguments{
\item{m}{An integer representing the value used to generate the factors. It must be a multiple of 2.}
}
\value{
A list of factors in the form of Galois Field elements or infinity elements. Each factor is a GFElem object or PGInf object, representing elements in the Galois Field or infinity respectively.
}
\description{
This function generates factors for a specific value of \eqn{q}, where \eqn{q = 2}. It calculates factors based on the Galois Field and infinity elements.
}
\examples{
# Generate factors for q = 2 with m = 6
factors <- Gen2Factors(6)
factors

}
