% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Gen3Factors}
\alias{Gen3Factors}
\title{Generate Factors for \eqn{q = 3}}
\usage{
Gen3Factors(m)
}
\arguments{
\item{m}{An integer representing the parameter for the factors. It must be a
multiple of 6, and \eqn{m-1} must be a prime number.}
}
\value{
A list of factors in the Galois Field GF(m-1), including the point at infinity.
        The result is unlisted and contains the generated factors.
}
\description{
This function generates factors for a given parameter \eqn{m} where \eqn{q = 3}.
The factors are generated using elements of the Galois Field GF(m-1) and
include the point at infinity. The function performs several checks to ensure
the validity of the input and uses parallel processing to compute the factors.
}
\details{
The function checks if \eqn{m} is a multiple of 6 and if \eqn{m-1} is a prime number.
It then calculates the factors using elements of the Galois Field GF(m-1) and
processes them using parallel computation with `future_map` to speed up the operation.
The factors include all elements of the field and are calculated using the
iterated function approach. The result is unlisted to return a simple list of factors.
}
\examples{
# Generate factors for q = 3
factors <- Gen3Factors(12)
factors
}
