% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PGField.R
\name{GFElem}
\alias{GFElem}
\title{Galois Field Element Class}
\description{
The `GFElem` class represents an element in a Galois Field (GF). A Galois Field, also known as a finite field, is a field with a finite number of elements. This class provides methods for initializing a field element, checking if it is zero, and displaying its value.
}
\details{
The `GFElem` class has two main fields:
\itemize{
  \item `value`: An integer that represents the value of the Galois field element.
  \item `p`: A prime number that represents the characteristic of the Galois field.
}

The class provides methods to:
\itemize{
  \item Initialize a Galois Field element with a specified value and characteristic.
  \item Check if the element is zero.
  \item Display the element in a readable format.
}
}
\examples{
# Create a Galois Field element
elem <- GFElem$new(3, 7)

# Display the Galois Field element
elem$show()

# Check if the element is zero
elem$iszero()

# Create a zero element and check
zero_elem <- GFElem$new(0, 7)
zero_elem$iszero()

# Create an element and check if it is zero
non_zero_elem <- GFElem$new(5, 7)
non_zero_elem$iszero()

# Display an element
non_zero_elem$show()
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{value}}{An integer representing the value of the Galois field element.}

\item{\code{p}}{A prime number representing the characteristic of the Galois field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GFElem-new}{\code{GFElem$new()}}
\item \href{#method-GFElem-iszero}{\code{GFElem$iszero()}}
\item \href{#method-GFElem-show}{\code{GFElem$show()}}
\item \href{#method-GFElem-clone}{\code{GFElem$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GFElem-new"></a>}}
\if{latex}{\out{\hypertarget{method-GFElem-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new GFElem object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GFElem$new(value, p)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{An integer. The value of the Galois field element.}

\item{\code{p}}{A prime number. The characteristic of the Galois field.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `GFElem` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GFElem-iszero"></a>}}
\if{latex}{\out{\hypertarget{method-GFElem-iszero}{}}}
\subsection{Method \code{iszero()}}{
Check if the element is zero.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GFElem$iszero()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TRUE if the element is zero, otherwise FALSE.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GFElem-show"></a>}}
\if{latex}{\out{\hypertarget{method-GFElem-show}{}}}
\subsection{Method \code{show()}}{
Display the element.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GFElem$show()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GFElem-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GFElem-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GFElem$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
