% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{HyperDesign}
\alias{HyperDesign}
\title{Generate a HyperDesign Matrix}
\usage{
HyperDesign(n, m, q)
}
\arguments{
\item{n}{An integer specifying the number of columns (combinations) in the matrix.}

\item{m}{An integer specifying the total number of elements to choose from.}

\item{q}{An integer specifying the number of elements to choose for each combination.}
}
\value{
A sparse matrix of class `CsparseMatrix` from the `Matrix` package, where
each column represents a combination of \eqn{q} elements from the set of \eqn{m} elements.
}
\description{
The `HyperDesign` function generates a design matrix for combinatorial designs.
Each column in the matrix represents a combination of \eqn{q} elements chosen
from a set of \eqn{m} elements. The resulting matrix is sparse.
}
\details{
The function generates all possible combinations of \eqn{q} elements from \eqn{m} elements.
It then selects \eqn{n} combinations and creates a matrix where each column represents
one of these combinations. The matrix is sparse, with 1s indicating the presence of
an element in the combination and 0s otherwise.

The function assumes that \eqn{m} is a multiple of 2 and \eqn{q} is a positive integer.
If the number of required combinations \eqn{n} exceeds the total number of possible combinations,
an error is thrown.
}
\examples{
# Generate a HyperDesign matrix with 10 columns, 6 elements, and combinations of 2
design1 <- HyperDesign(n = 10, m = 6, q = 2)
design1

# Generate a HyperDesign matrix with 12 columns, 6 elements, and combinations of 2
design2 <- HyperDesign(n = 12, m = 6, q = 2)
design2

# Generate a HyperDesign matrix with 10 columns, 6 elements, and combinations of 3
design3 <- HyperDesign(n = 10, m = 6, q = 3)
design3
}
\seealso{
The `Matrix` package for sparse matrix operations.
}
