% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Iterated}
\alias{Iterated}
\title{Iterated Function Application}
\usage{
Iterated(f, x)
}
\arguments{
\item{f}{A function to apply iteratively. It should take a single argument and return a result of the same type as the input.}

\item{x}{An initial value to start the iteration. It should be of the same type as the input required by the function `f`.}
}
\value{
A list containing the sequence of values obtained by applying the function `f` iteratively to `x`.
The sequence ends when a value repeats, and the repeated value is included in the result.
}
\description{
Applies a function iteratively to an initial element until the result repeats.
This function generates a sequence of values starting from an initial value,
where each subsequent value is obtained by applying a given function to the previous value.
The process stops when a repeated value is encountered.
}
\examples{
# Define a function to be iterated
multiply_by_two <- function(x) x * 2

# Apply the function iteratively
result <- Iterated(multiply_by_two, 1)
print(result)

# Define a GFElem element and apply a function to it
elem <- GFElem$new(3, 7)
iterated <- Iterated(function(x) Multiply(x, elem), elem)
iterated
}
