% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HyperGen.R
\name{Multiply}
\alias{Multiply}
\title{Custom Multiplication Function}
\usage{
Multiply(x, y)
}
\arguments{
\item{x}{An object of class `GFElem` or `PGInf`. Represents the first operand in the multiplication.}

\item{y}{An object of class `GFElem` or `PGInf`. Represents the second operand in the multiplication.}
}
\value{
An object of class `GFElem` or `PGInf`, depending on the types of the input arguments.
If both inputs are `GFElem`, the result is a new `GFElem` with the product of their values.
If either input is `PGInf`, the result is `PGInf` with the same parameter as the inputs.
}
\description{
This function multiplies two objects, which can either be elements of a Galois Field (GFElem)
or elements representing infinity (PGInf). If either argument is of type `PGInf`, the function
handles special cases where the multiplication involves a zero element in the Galois Field.
}
\examples{
# Multiply two Galois Field elements
elem1 <- GFElem$new(3, 7)
elem2 <- GFElem$new(5, 7)
result <- Multiply(elem1, elem2)
result$show()

# Multiply a Galois Field element by infinity
elem <- GFElem$new(4, 7)
inf <- PGInf$new(7)
result_inf <- Multiply(elem, inf)
result_inf$show()

# Multiply two infinity elements
inf1 <- PGInf$new(7)
inf2 <- PGInf$new(7)
result_inf <- Multiply(inf1, inf2)
result_inf$show()

}
