% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HyperGen.R
\name{Negate}
\alias{Negate}
\title{Custom Negation Function}
\usage{
Negate(x)
}
\arguments{
\item{x}{An object of class \code{GFElem} or \code{PGInf}. For \code{GFElem}, it should be a Galois Field element. For \code{PGInf}, it should be a projective infinity element.}
}
\value{
For \code{GFElem}, returns a new \code{GFElem} object with the negated value. For \code{PGInf}, returns a new \code{PGInf} object with the same parameter.
}
\description{
Negates a Galois Field element or a projective infinity element.
If the input is a Galois Field element, it returns a new Galois Field element with the negated value.
If the input is a projective infinity element, it returns a new projective infinity element with the same parameter.
}
\examples{
# Negate a Galois Field element
elem <- GFElem$new(3, 7)
result <- Negate(elem)
result$show()

# Negate a projective infinity element
inf_elem <- PGInf$new(7)
result_inf <- Negate(inf_elem)
result_inf$show()

}
