% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HyperGen.R
\name{One}
\alias{One}
\title{Generate the One Element of a Galois Field}
\usage{
One(x)
}
\arguments{
\item{x}{An object of class `GFElem` or `PGInf`. The function uses the characteristic `p` of the input element to create a `GFElem` object with value `1`.}
}
\value{
An object of class `GFElem` representing the element `1` in the Galois Field defined by the characteristic `p` of the input element.
}
\description{
This function returns the element `1` in the context of a Galois Field or a projective geometry setting. The returned element is always a `GFElem` object with value `1` and the same characteristic as the input element.
}
\examples{
# Create a GFElem object with value 3 and characteristic 7
elem <- GFElem$new(3, 7)

# Generate the one element of the same Galois Field
one_elem <- One(elem)

# Display the one element
one_elem$show()

# Create a PGInf object with characteristic 7
pg_inf <- PGInf$new(7)

# Generate the one element for PGInf (should be the same GFElem)
one_pg_inf <- One(pg_inf)
one_pg_inf$show()

}
