% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Order}
\alias{Order}
\title{Calculate the Order of a Galois Field Element}
\usage{
Order(x)
}
\arguments{
\item{x}{An object of class `GFElem` representing the Galois Field element whose order is to be computed.}
}
\value{
An integer representing the order of the Galois Field element.
}
\description{
This function calculates the order of a Galois Field (GF) element. The order of an element in a finite field is the smallest positive integer \eqn{n} such that the element raised to the power \eqn{n} is equal to 1. For elements in the field GF(p), where \eqn{p} is a prime number, the order is a divisor of \eqn{p-1}.
}
\examples{
# Compute the order of a Galois Field element
elem <- GFElem$new(3, 7)
order <- Order(elem)
print(order)

# Example with a different element
elem2 <- GFElem$new(2, 11)
order2 <- Order(elem2)
print(order2)

# Order of zero element
zero_elem <- GFElem$new(0, 5)
order_zero <- Order(zero_elem)
print(order_zero)

}
