% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StabilityApp.R
\name{StabilityApp}
\alias{StabilityApp}
\title{StabilityApp}
\value{
output of \link[stability]{stability} R package
}
\description{
StabilityApp
}
\examples{
if(interactive()) {
    library(StabilityApp)
    StabilityApp()
  }

}
