% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HyperGen.R
\name{Zero}
\alias{Zero}
\title{Zero Function}
\usage{
Zero(x)
}
\arguments{
\item{x}{An object of class `GFElem` or `PGInf`. The field order (prime) of the input element will be used
to create the zero element.}
}
\value{
An object of class `GFElem` with value 0 and the same field order as the input element.
}
\description{
Creates a zero element in the Galois Field or Projective Geometry.
}
\details{
This function returns a zero element corresponding to the input element's field or space.
For `GFElem` objects, it returns a `GFElem` with value 0 and the same field order as the input.
For `PGInf` objects, it returns a `GFElem` with value 0 and the same field order as the input.
}
\examples{
# Create the zero element of a Galois Field
elem <- GFElem$new(3, 7)
zero_elem <- Zero(elem)
zero_elem$show()

# Create the zero element of Projective Geometry
pg_inf <- PGInf$new(7)
zero_pg <- Zero(pg_inf)
zero_pg$show()

}
